chore: move modal steps to their own component folder
